backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
alembic
boto3
reportlab
